Refactored expression matching into Expression
