+++ lib/environment.js
@@ -1,3 +1,5 @@
+var Expression = require('./expression');
+
 var cache = {};
 
 var Environment = module.exports = function(jsonObject, expression, options) {
@@ -114,71 +116,5 @@ Environment.prototype.slice = function(segment, expression, object, path) {
 Environment.prototype.match = function(expression, object, _vname) {
   if(object === null) return object;
 
-  if(/[>=<]/.test(expression))
-    return this.compareValue(expression, object);
-  else if(/[-+*\/]/.test(expression))
-    return this.computeValue(expression, object);
-  else if(/\./.test(expression))
-    return this.getProperty(expression, object);
-  else
-    throw new SyntaxError("jsonPath: invalid query syntax: " + expression);
-};
-
-Environment.prototype.compareValue = function(expression, object) {
-  return this.operate(object, expression, '>=<');
-};
-
-Environment.prototype.getProperty = function(expression, object) {
-  var parts = expression.match(/\.([^\s]+)/);
-  var rhs;
-  try {
-    rhs = JSON.parse('"' + parts[1] + '"');
-  }
-  catch(e) {
-    throw new SyntaxError("jsonPath: invalid property name: " + parts[1]);
-  }
-  return object[rhs];
-};
-
-Environment.prototype.getValue = function(expression) {
-  try {
-    return JSON.parse(expression);
-  }
-  catch(e) {
-    throw new SyntaxError("jsonPath: invalid right-hand-side value: " + expression);
-  }
-};
-
-Environment.prototype.computeValue = function(expression, object) {
-  return this.operate(object, expression, '-+*/');
-};
-
-Environment.prototype.operate = function(object, expression, operators) {
-  var regex = new RegExp("([^\(\)" + operators + "]+)([" + operators + "]+)([^\(\)]+)");
-  var parts = expression.match(regex);
-  var lhs = this.getProperty(parts[1], object);
-  var rhs = this.getValue(parts[3]);
-
-  switch(parts[2]) {
-  case '==':
-    return lhs == rhs;
-  case '>=':
-    return lhs >= rhs;
-  case '<=':
-    return lhs <= rhs;
-  case '>':
-    return lhs > rhs;
-  case '<':
-    return lhs < rhs;
-  case '-':
-    return lhs - rhs;
-  case '+':
-    return lhs + rhs;
-  case '*':
-    return lhs * rhs;
-  case '/':
-    return lhs / rhs;
-  default:
-    throw new SyntaxError("jsonPath: invalid query syntax: " + expression);
-  }
+  return Expression.execute(expression, object);
 };
